# client-api/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Matrix Client-Server Api",
    "version" : "r0.3.0",
    "title" : "Matrix Client-Server Api",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Anatoly",
      "url" : "https://github.com/ma1uta/jeon",
      "email" : "sablintolya@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "matrix.org:8448",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/_matrix/client/r0/profile/{userId}" : {
      "get" : {
        "tags" : [ "Profile" ],
        "summary" : "Get the combined profile information for this user. This API may be used to fetch the user's own profile information or other users; either locally or on remote homeservers. This API may return keys which are not limited to displayname or avatar_url.",
        "description" : "",
        "operationId" : "profile",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The user whose profile information to get.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "The avatar URL for this user.",
            "schema" : {
              "$ref" : "#/definitions/Profile"
            }
          },
          "404" : {
            "description" : "There is no profile information for this user or this user does not exist."
          }
        }
      }
    },
...
    "/_matrix/client/r0/rooms/{roomId}/context/{eventId}" : {
      "get" : {
        "tags" : [ "EventContext" ],
        "summary" : "This API returns a number of events that happened just before and after the specified event.",
        "description" : "This allows clients to get the context surrounding an event.",
        "operationId" : "context",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "description" : "The room to get events from.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "eventId",
          "in" : "path",
          "description" : "The event to get context around.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The maximum number of events to return. Default: 10.",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "The events and state surrounding the requested event.",
            "schema" : {
              "$ref" : "#/definitions/EventContextResponse"
            }
          }
        }
      }
    },
...
    "/_matrix/client/r0/rooms/{roomId}/messages" : {
      "get" : {
        "tags" : [ "Event" ],
        "summary" : "This API returns a list of message and state events for a room.",
        "description" : "It uses pagination query parameters to paginate history in the room.",
        "operationId" : "messages",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "description" : "The room to get events from.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "from",
          "in" : "query",
          "description" : "The token to start returning events from. This token can be obtained from a prev_batch token returned for each room by the sync API, or from a start or end token returned by a previous request to this endpoint.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "to",
          "in" : "query",
          "description" : "The token to stop returning events at. This token can be obtained from a prev_batch token returned for each room by the sync endpoint, or from a start or end token returned by a previous request to this endpoint.",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "dir",
          "in" : "query",
          "description" : "The direction to return events from.",
          "required" : true,
          "type" : "string",
          "enum" : [ "['b'", "'f]" ]
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The maximum number of events to return. Default: 10.",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "filter",
          "in" : "query",
          "description" : "A JSON RoomEventFilter to filter returned events with.",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of messages with a new token to request more.",
            "schema" : {
              "$ref" : "#/definitions/Page"
            }
          },
          "403" : {
            "description" : "You aren't a member of the room."
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Event" : {
      "type" : "object",
      "required" : [ "event_id", "origin_server_ts", "room_id", "sender", "state_key", "type" ],
      "properties" : {
        "content" : {
          "description" : "The fields in this object will vary depending on the type of event. When interacting with the REST API, this is the HTTP body.",
          "$ref" : "#/definitions/EventContent"
        },
        "type" : {
          "type" : "string",
          "description" : "The type of event. This SHOULD be namespaced similar to Java package naming conventions."
        },
        "sender" : {
          "type" : "string",
          "description" : "Contains the fully-qualified ID of the user who sent this event."
        },
        "unsigned" : {
          "description" : "Contains optional extra information about the event.",
          "$ref" : "#/definitions/Unsigned"
        },
        "event_id" : {
          "type" : "string",
          "description" : "The globally unique event identifier."
        },
        "room_id" : {
          "type" : "string",
          "description" : "The ID of the room associated with this event."
        },
        "origin_server_ts" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Timestamp in milliseconds on originating homeserver when this event was sent."
        },
        "prev_content" : {
          "description" : "The previous content for this event. If there is no previous content, this key will be missing.",
          "$ref" : "#/definitions/EventContent"
        },
        "state_key" : {
          "type" : "string",
          "description" : " A unique key which defines the overwriting semantics for this piece of room state. This value is often a zero-length string. The presence of this key makes this event a State Event. The key MUST NOT start with '_'."
        },
        "invite_room_state" : {
          "type" : "array",
          "description" : "This contains an array of StrippedState Events. These events provide information on a subset of state events such as the room name.",
          "items" : {
            "$ref" : "#/definitions/StrippedState"
          }
        }
      },
      "description" : "Event."
    },
    "EventContent" : {
      "type" : "object",
      "discriminator" : "type",
      "description" : "Parent class of all event contents part in the Event."
    },
...
    "EventContextResponse" : {
      "type" : "object",
      "properties" : {
        "start" : {
          "type" : "string",
          "description" : "Pagination token for the start of the chunk."
        },
        "end" : {
          "type" : "string",
          "description" : "Pagination token for the end of the chunk."
        },
        "profile_info" : {
          "type" : "object",
          "description" : "The historic profile information of the users that sent the events returned.",
          "additionalProperties" : {
            "$ref" : "#/definitions/Profile"
          }
        },
        "events_before" : {
          "type" : "array",
          "description" : "Events just before the result.",
          "items" : {
            "$ref" : "#/definitions/Event"
          }
        },
        "events_after" : {
          "type" : "array",
          "description" : "Events just after the result.",
          "items" : {
            "$ref" : "#/definitions/Event"
          }
        }
      },
      "description" : "Event context."
    },
...
    "Page" : {
      "type" : "object",
      "properties" : {
        "start" : {
          "type" : "string",
          "description" : "The token the pagination starts from. If dir=b this will be the token supplied in from."
        },
        "end" : {
          "type" : "string",
          "description" : "The token the pagination ends at. If dir=b this token should be used again to request even earlier events."
        },
        "chunk" : {
          "type" : "array",
          "description" : "A list of room events.",
          "items" : {
            "type" : "object"
          }
        }
      },
      "description" : "Pagination."
    },
...
    "Profile" : {
      "type" : "object",
      "properties" : {
        "displayname" : {
          "type" : "string",
          "description" : "The user's display name if they have set one, otherwise not present."
        },
        "avatar_url" : {
          "type" : "string",
          "description" : "The user's avatar URL if they have set one, otherwise not present."
        }
      },
      "description" : "JSON body for profile api (profile)."
    },
...
    "StrippedState" : {
      "type" : "object",
      "required" : [ "content", "state_key", "type" ],
      "properties" : {
        "content" : {
          "type" : "object",
          "description" : "The content for the event.",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "type" : {
          "type" : "string",
          "description" : "The type for the event."
        },
        "state_key" : {
          "type" : "string",
          "description" : "The state_key for the event."
        }
      },
      "description" : "Provides information on a subset of state events such as the room name."
    },
...
    "Unsigned" : {
      "type" : "object",
      "properties" : {
        "age" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "The time in milliseconds that has elapsed since the event was sent. This field is generated by the local homeserver, and may be incorrect if the local time on at least one of the two servers is out of sync, which can cause the age to either be negative or greater than it actually is."
        },
        "redacted_because" : {
          "description" : "The event that redacted this event, if any.",
          "$ref" : "#/definitions/Event"
        },
        "transaction_id" : {
          "type" : "string",
          "description" : "The client-supplied transaction ID, if the client being given the event is the same one which sent it."
        },
        "prev_content" : {
          "type" : "object",
          "description" : "The previous content for this state. This will be present only for state events appearing in the timeline. If this is not a state event, or there is no previous content, this key will be missing.",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      },
      "description" : "Unsigned addition data."
    },
...
  }
}